requirements.txt
cognite-sdk
pandas
pydantic
pydantic-settings
xlsxwriter
